+++ src/backend/backends.rs
@@ -1,7 +1,7 @@
-use crate::data_types::OmFileArrayDataType;
-use crate::om::c_defaults::{new_data_read, new_index_read};
-use crate::om::errors::OmFilesRsError;
-use crate::om::mmapfile::{MAdvice, MmapFile, MmapType};
+use crate::backend::mmapfile::{MAdvice, MmapFile, MmapType};
+use crate::core::c_defaults::{new_data_read, new_index_read};
+use crate::core::data_types::OmFileArrayDataType;
+use crate::errors::OmFilesRsError;
 use omfileformatc_rs::{
     om_decoder_decode_chunks, om_decoder_next_data_read, om_decoder_next_index_read, OmDecoder_t,
     OmError_t_ERROR_OK,
