+++ src/bin/reformat.rs
@@ -1,6 +1,6 @@
-use omfiles_rs::compression::CompressionType;
-use omfiles_rs::om::reader::OmFileReader;
-use omfiles_rs::om::writer::OmFileWriter;
+use omfiles_rs::core::compression::CompressionType;
+use omfiles_rs::io::reader::OmFileReader;
+use omfiles_rs::io::writer::OmFileWriter;
 use std::{
     io::{self},
     rc::Rc,
