+++ src/bin/testread.rs
@@ -1,4 +1,4 @@
-use omfiles_rs::om::reader::OmFileReader;
+use omfiles_rs::io::reader::OmFileReader;
 use std::env;
 use std::io::{self};
 use std::ops::Range;
