+++ src/core/c_defaults.rs
