+++ src/io/aligned_buffer.rs
