+++ src/io/reader.rs
@@ -1,12 +1,12 @@
-use crate::aligned_buffer::as_typed_slice;
-use crate::compression::{p4ndec256_bound, CompressionType};
-use crate::om::backends::OmFileReaderBackend;
-use crate::om::c_defaults::create_decoder;
-use crate::om::dimensions::Dimensions;
-use crate::om::errors::OmFilesRsError;
-use crate::om::header::OmHeader;
-use crate::om::mmapfile::{MmapFile, Mode};
-use crate::om::reader2::OmFileReader2;
+use crate::backend::backends::OmFileReaderBackend;
+use crate::backend::mmapfile::{MmapFile, Mode};
+use crate::core::c_defaults::create_decoder;
+use crate::core::compression::{p4ndec256_bound, CompressionType};
+use crate::core::dimensions::Dimensions;
+use crate::core::header::OmHeader;
+use crate::errors::OmFilesRsError;
+use crate::io::aligned_buffer::as_typed_slice;
+use crate::io::reader2::OmFileReader2;
 use crate::utils::{add_range, divide_range};
 use omfileformatc_rs::{om_decoder_init, OmError_t_ERROR_OK};
 use std::fs::File;
