--- src/om/mod.rs
@@ -1,11 +1,0 @@
-pub mod backends;
-mod buffered_writer;
-mod c_defaults;
-mod dimensions;
-pub mod errors;
-mod header;
-pub mod mmapfile;
-pub mod reader;
-pub mod reader2;
-pub mod writer;
-pub mod writer2;
